show room messages
