+++ src/components/chat/ChatPage.js
@@ -1,8 +1,11 @@
 import React, { Component, PropTypes }  from 'react';
 import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
+import * as firebase from 'firebase/firebase-browser';
 
 import checkAuth from '../requireAuth';
+import {getUsersSuccess} from '../../actions/chatActions';
+
 import MessagesThread from './MessagesThread.js';
 import MessagesInput from './MessagesInput.js';
 import RoomSelection from './RoomSelection.js';
@@ -10,6 +13,13 @@ import RoomCreation from './RoomCreation.js';
 import OrSeparator from './OrSeparator.js';
 
 class ChatPage extends Component {
+  componentWillMount() {
+    const usersRef = firebase.database().ref('users');
+    usersRef.on('child_added', data => {
+      this.props.getUsersSuccess(data.key, data.val());
+    });
+  }
+
   render() {
     return !this.props.currentRoom
       ? (
@@ -24,13 +34,16 @@ class ChatPage extends Component {
       ) : (
       <div>
         <h1>Chat - Discuss</h1>
+        <MessagesThread />
+        <MessagesInput />
       </div>
     );
   }
 }
 
 ChatPage.propTypes =  {
-  currentRoom: PropTypes.string
+  currentRoom: PropTypes.string,
+  getUsersSuccess: PropTypes.func
 };
 
 function mapStateToProps(state) {
@@ -39,4 +52,10 @@ function mapStateToProps(state) {
   };
 }
 
-export default connect(mapStateToProps)(checkAuth(ChatPage));
+function mapDispatchToProps(dispatch) {
+  return {
+    getUsersSuccess: bindActionCreators(getUsersSuccess, dispatch)
+  };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(checkAuth(ChatPage));
