+++ src/components/chat/RoomCreation.js
@@ -20,7 +20,8 @@ class RoomCreation extends Component {
     event.preventDefault();
     if (this.props.temporaryRoom) {
       this.props.createRoom({
-        name: this.props.temporaryRoom
+        name: this.props.temporaryRoom,
+        messages: {}
       });
     }
   }
